+++ extensions/himalaya/src/envelopes.ts
@@ -1,12 +1,15 @@
 import { Envelope, Flag } from "./models";
 import * as Exec from "./exec";
 
-export async function list(folder: string): Promise<Envelope[]> {
-  const { stdout, stderr } = await Exec.run(`himalaya --folder "${folder}" -o json list -s 100`, {
-    env: {
-      PATH: Exec.PATH,
-    },
-  });
+export async function list(folder: string, account: string): Promise<Envelope[]> {
+  const { stdout, stderr } = await Exec.run(
+    `himalaya --account "${account}" --folder "${folder}" -o json list -s 100`,
+    {
+      env: {
+        PATH: Exec.PATH,
+      },
+    }
+  );
 
   if (stdout) {
     const results: any = JSON.parse(stdout);