+++ extensions/himalaya/src/state.ts
@@ -4,6 +4,7 @@ export interface State {
   isLoading: boolean;
   envelopes: Envelope[];
   folders: Folder[];
-  currentFolderName: string;
   exe: boolean;
+  currentFolderName: string;
+  currentAccountName: string;
 }