+++ extensions/himalaya/CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Folder support for the Index command
+- Folder support from search bar and preferences
+- Account support from preferences
 
 ## [0.2.0] - 2023-04-28